requirements.txt
alembic==1.13.2
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
PyMySQL==1.1.1
python-dotenv==1.0.1
SQLAlchemy==2.0.43
typing_extensions==4.12.2
Werkzeug==3.0.4
